inventory_api/python/requirements.txt
hvac
flask
paramiko
psycopg2-binary